Ensure setup.py is removed when building

+++ scripts/build.sh
@@ -12,7 +12,11 @@ cd ../client
 # As this is an x86_64 only version, ensure `py2app` outputs x86_64.
 python3 -m pip install -r requirements.txt pyqt6 py2app GitPython
 python3 _version.py
-rm setup.py
+
+# Recreate our setup.py with macOS-specific options.
+if [ -f setup.py ]; then
+  rm setup.py
+fi
 py2applet --make-setup app.py icon.icns "icon.png" "taskbarDark.png" "taskbarLight.png" "version.txt"
 sed -i '' -e "s/)/    name='NSO-RPC')/" setup.py
 python3 setup.py py2app --arch=x86_64 -O2
