+++ scripts/macos-universal2/build.sh
@@ -15,11 +15,14 @@ bash prep-PyQt.sh
 python3 -m pip install PyQt6
 
 # Lastly, build our client.
-# We'll override our default setup.py with one adjusted for macOS.
 cd ../../client
 python3 -m pip install -r requirements.txt py2app GitPython
 python3 _version.py
-rm setup.py
+
+# Recreate our setup.py with macOS-specific options.
+if [ -f setup.py ]; then
+  rm setup.py
+fi
 py2applet --make-setup app.py icon.icns "icon.png" "taskbarDark.png" "taskbarLight.png" "version.txt"
 # build universal binary
 sed -i '' -e "s/)/    name='NSO-RPC')/" setup.py
